Add Tests

+++ Imparta.Application/Task/TaskListService.cs
@@ -18,6 +18,8 @@ namespace Imparta.Application.Task
 
         public async Task<TaskList> AddAsync(TaskList taskList)
         {
+            if (taskList == null) throw new ArgumentNullException();
+
             return await _taskListRepo.AddAsync(taskList);
         }
 
@@ -28,26 +30,36 @@ namespace Imparta.Application.Task
 
         public async Task<IList<TaskList>> GetAllAsync(string userId)
         {
+            if (string.IsNullOrWhiteSpace(userId)) throw new ArgumentNullException();
+
             return await _taskListRepo.GetListsForUser(userId);
         }
 
         public async Task<TaskList> GetByIdAync(Guid id)
         {
+            if (id == null || id == Guid.Empty) throw new ArgumentNullException();
+
             return await _taskListRepo.GetByIdAsync(id);
         }
 
         public async Task<IList<TaskModel>> GetTasksForListWithId(Guid id)
         {
+            if (id == null || id == Guid.Empty) throw new ArgumentNullException();
+
             return await _taskListRepo.GetTasks(id);
         }
 
         public async System.Threading.Tasks.Task RemoveAsync(Guid id)
         {
+            if (id == null || id == Guid.Empty) throw new ArgumentNullException();
+
             await _taskListRepo.DeleteAsync(id);
         }
 
         public async Task<TaskList> UpdateAsync(TaskList taskList)
         {
+            if (taskList == null) throw new ArgumentNullException();
+
             await _taskListRepo.UpdateAsync(taskList);
             return taskList;
         }
